railfence: fix usages of unsigned and size_t
'size_t' is used to represent size of objects.
'unsigned' seems fit for arithmetic involving non-negative
numbers.

So, fix the usages by using unsigned and size_t where
appropriate.

Resources:

https://www.embedded.com/electronics-blogs/programming-pointers/4026076/Why-size-t-matters#

+++ railfence/railfence.cpp
@@ -15,7 +15,7 @@ namespace railfence {
 		size_t rows, cols;
 
 		public:
-			railfence_matrix(unsigned length, unsigned cols, bool encrypt) {
+			railfence_matrix(size_t length, size_t cols, bool encrypt) {
 				this-> rows = length/cols;
 				this->cols = cols;
